requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
jupyter
sqlalchemy
psycopg2
python-dotenv